requirements.txt
requests==2.25.1
SQLAlchemy-Utils==0.36.8
pytest==6.2.1